requirements/dev.txt
celery
django
django-celery-email
django-environ
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
django-autoslug
django-countries
django-phonenumber-field
drf-yasg
flower
importlib-metadata
pillow
psycopg2-binary
phonenumbers
PyJWT
pytz
redis